* commit '1c3b11a4563b75e3bb8f0c084553e1c844ee368e': (85 commits)
  remove magic number #QAB-190
  Update use of the margin #QAB-188
  Try Transform after exploit #QAB-189
  Refactoring resources and use of instanceOf #QAB-187
  Decrease amount even if the transforme fail (even production == 0 ) #QAB-176 #QAB-155
  Change margin error to transform instead of recipe #QAB-188
  Add response test to transform #QAB-169
  Add test to check return statet to Transform #QAB-169
  Fix enoughToTransform() to use the contracts #QAB-115, add test to veiry scout retourn in Transform #QAB-169
  Delete tests failed #QAB-167
  Stop using auxiliar of manufacturedResources and use already the contract #QAB-115
  Change method equals, change class to abstract #QAB-187 #QAB-70
  less debt #QAB-157
  Fix recipe of INGOT #QAB-186
  Change attribut of list related to manufactured contracts of class #QAB-181
  Ignore test to fix #QAB-167
  necessary changes due to merge #QAB-181
  test ignore to merge with change in the other branch related to #QAB-181
  Forgot file of commit be02be0158c #QAB-176 #QAB-115
  Benchmark03, fix error and add test.
  ...

+++ src/main/java/fr/unice/polytech/qgl/qab/map/MapHandler.java
@@ -32,17 +32,17 @@ public class MapHandler {
         return Math.abs(p.getY() - lastPosition.getY());
     }
 
-    public static double getDistance(Position current, Position p) {
-        double distX = getDistX(current, p);
-        double distY = getDistY(current, p);
+    public static int getDistance(Position current, Position p) {
+        int distX = getDistX(current, p);
+        int distY = getDistY(current, p);
         return distX + distY;
     }
 
-    public static double getDistY(Position current, Position p) {
-        return (double) Math.abs(p.getY() - current.getY());
+    public static int getDistY(Position current, Position p) {
+        return Math.abs(p.getY() - current.getY());
     }
 
-    public static double getDistX(Position current, Position p) {
-        return (double) Math.abs(p.getX() - current.getX());
+    public static int getDistX(Position current, Position p) {
+        return Math.abs(p.getX() - current.getX());
     }
 }
